Satisfy CI lint: immutable NamedTuple default, drop a no-op int() and an unused binding
pic_specs' class-level default becomes a tuple — a shared mutable default on
a NamedTuple is exactly what RUF012 exists to catch, and the field is never
mutated in place by contract anyway. round() already returns an int (RUF046),
and the image-jobs workflow test bound its env() to a name nothing read.

+++ backend/app/avatar_pool.py
@@ -503,7 +503,7 @@ def _face_pct(v) -> int | None:
     except (TypeError, ValueError):
         return None
     if f <= 1.0:
-        return int(round(f * 100))
+        return round(f * 100)
     return int(f)
